Fix compilation on linux

+++ PDS/3dEngine_flush.cpp
@@ -4,11 +4,11 @@
 #include <OpenGLES/ES3/gl.h>
 #include <OpenGLES/ES3/glext.h>
 #else
-#include <gl/gl3w.h>
+#include <GL/gl3w.h>
 #ifdef __MACOS__
 #include <Opengl/gl.h>
 #else
-#include <gl/gl.h>
+#include <GL/gl.h>
 #endif
 #endif
 
@@ -239,7 +239,7 @@ void GetDistanceFalloff(float* falloutColor, s_objectToRender* pObject, int vert
 
     colorIndex /= 128;
 
-    colorIndex -= fmod(colorIndex, 8);
+    colorIndex -= fmodf(colorIndex, 8);
     colorIndex /= 6;
 
     int value = (int)colorIndex;
